+++ esp-hal/src/soc/esp32c2/clocks.rs
@@ -18,6 +18,7 @@ use esp_rom_sys::rom::{ets_delay_us, ets_update_cpu_frequency_rom};
 
 use crate::{
     peripherals::{I2C_ANA_MST, LPWR, SYSTEM, TIMG0},
+    rtc_cntl::Rtc,
     soc::regi2c,
     time::Rate,
 };
@@ -184,16 +185,15 @@ fn configure_xtal_clk_impl(_clocks: &mut ClockTree, config: XtalClkConfig) {
     // similar to ESP-IDF, just in case something relies on that, or, if we can in the future read
     // back the value instead of wasting RAM on it.
 
-    const DISABLE_ROM_LOG: u32 = 1;
-
+    // Used by `rtc_clk_xtal_freq_get` patched ROM function.
     let freq_mhz = config.value() / 1_000_000;
     LPWR::regs().store4().modify(|r, w| unsafe {
         // The data is stored in two copies of 16-bit values. The first bit overwrites the LSB of
         // the frequency value with DISABLE_ROM_LOG.
 
         // Copy the DISABLE_ROM_LOG bit
-        let disable_rom_log_bit = r.bits() & DISABLE_ROM_LOG;
-        let half = (freq_mhz & (0xFFFF & !DISABLE_ROM_LOG)) | disable_rom_log_bit;
+        let disable_rom_log_bit = r.bits() & Rtc::RTC_DISABLE_ROM_LOG;
+        let half = (freq_mhz & (0xFFFF & !Rtc::RTC_DISABLE_ROM_LOG)) | disable_rom_log_bit;
         w.data().bits(half | (half << 16))
     });
 }