+++ esp-metadata-generated/src/_generated_esp32c2.rs
@@ -54,9 +54,6 @@ macro_rules! property {
     ("soc.rc_slow_clock", str) => {
         stringify!(136000)
     };
-    ("soc.has_multiple_xtal_options") => {
-        true
-    };
     ("assist_debug.has_sp_monitor") => {
         true
     };
@@ -203,15 +200,6 @@ macro_rules! property {
     };
     ("phy.combo_module") => {
         true
-    };
-}
-#[macro_export]
-#[cfg_attr(docsrs, doc(cfg(feature = "_device-selected")))]
-macro_rules! for_each_soc_xtal_options {
-    ($($pattern:tt => $code:tt;)*) => {
-        macro_rules! _for_each_inner { $(($pattern) => $code;)* ($other : tt) => {} }
-        _for_each_inner!((26)); _for_each_inner!((40)); _for_each_inner!((all(26),
-        (40)));
     };
 }
 #[macro_export]