+++ esp-metadata-generated/src/_generated_esp32h2.rs
@@ -54,12 +54,6 @@ macro_rules! property {
     ("soc.rc_slow_clock", str) => {
         stringify!(136000)
     };
-    ("soc.xtal_frequency") => {
-        32
-    };
-    ("soc.xtal_frequency", str) => {
-        stringify!(32)
-    };
     ("aes.dma") => {
         true
     };
@@ -263,14 +257,6 @@ macro_rules! property {
     };
     ("phy.combo_module") => {
         false
-    };
-}
-#[macro_export]
-#[cfg_attr(docsrs, doc(cfg(feature = "_device-selected")))]
-macro_rules! for_each_soc_xtal_options {
-    ($($pattern:tt => $code:tt;)*) => {
-        macro_rules! _for_each_inner { $(($pattern) => $code;)* ($other : tt) => {} }
-        _for_each_inner!((32)); _for_each_inner!((all(32)));
     };
 }
 #[macro_export]